Modify files and the repo name
