+++ requirements.txt
@@ -1,12 +1,7 @@
 flask==0.12
 python-dotenv==0.6.3
-Jinja2==2.8
-MarkupSafe==0.23
-Werkzeug==0.11.10
-click==6.6
 gunicorn==19.6.0
-itsdangerous==0.24
 requests==2.13.0
 flask-sqlalchemy==2.2
 Flask-Migrate==2.0.3
-gunicorn==19.7.1
+fbmq==2.0.1
